+++ edmu/models.py
@@ -1,4 +1,10 @@
-from django.contrib.auth.models import User
+try:
+    from django.contrib.auth import get_user_model
+except ImportError: # django < 1.5
+    from django.contrib.auth.models import User
+else:
+    User = get_user_model()
+
 from django.db import models
 from django.utils.translation import ugettext_lazy as _